+++ packages/caleuche-cli/src/index.ts
@@ -11,10 +11,17 @@ program
   .version(version);
 
 program
-  .command("compile <sample-directory> <data-file> <output-directory>")
-  .option("-p, --project", "Generate project file")
+  .command("compile")
+  .argument("<sample-path>", "Path to the sample file or the directory that containes it.")
+  .argument("<data-file>", "Path to the input to be used to compile the template.")
+  .argument("<output-directory>", "Directory where the compiled output will be written.")
+  .option("-p, --project", "Flag to indicate whether to generate the appropriate project file along with the compiled template.")
   .action(compile);
 
-program.command("batch <batch-file>").action(batchCompile);
+program
+  .command("batch")
+  .argument("<batch-file>", "Path to the batch file")
+  .option("-d, --output-dir <outputDir>", "Output directory for compiled samples")
+  .action(batchCompile);
 
 program.parse();