+++ packages/caleuche-cli/test/common.test.ts
@@ -5,28 +5,12 @@ vi.mock("fs");
 import fs from "fs";
 const mockFs = vi.mocked(fs);
 
-vi.mock("@caleuche/core");
-import { compileSample, Sample } from "@caleuche/core";
-const mockCompileSample = vi.mocked(compileSample);
-
 vi.mock("../src/utils");
-import {
-  parse,
-  resolveSampleFile,
-  createOutputDirectory,
-  resolveTemplate,
-  isVariantDefinition,
-  isVariantPath,
-  isVariantReference,
-} from "../src/utils";
+import { parse, resolveSampleFile, isFile } from "../src/utils";
 import { resolveAndParseSample } from "../src/common";
 const mockParse = vi.mocked(parse);
 const mockResolveSampleFile = vi.mocked(resolveSampleFile);
-const mockCreateOutputDirectory = vi.mocked(createOutputDirectory);
-const mockResolveTemplate = vi.mocked(resolveTemplate);
-const mockIsVariantDefinition = vi.mocked(isVariantDefinition);
-const mockIsVariantPath = vi.mocked(isVariantPath);
-const mockIsVariantReference = vi.mocked(isVariantReference);
+const mockIsFile = vi.mocked(isFile);
 
 describe("common", () => {
   describe("resolveAndParseSample", () => {
@@ -56,7 +40,7 @@ describe("common", () => {
 
     it("should return null and log an error when sample file cannot be parsed", () => {
       mockResolveSampleFile.mockReturnValue("/path/to/sample.yaml");
-      mockFs.existsSync.mockReturnValue(true);
+      mockIsFile.mockReturnValue(true);
       mockParse.mockReturnValueOnce(null);
 
       const sample = resolveAndParseSample("sample");